old gonio creation -> test; make generic util
